+++ springbootlearning-master/springBoot_ecommerceDemo/src/main/java/ecommerce/springBoot_ecommerceDemo/repository/CustomerRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,11 +17,12 @@ public interface CustomerRepository extends JpaRepository<Customer, Long>{
     // jpql query = find customers who spent more than specified amount
     @Query("SELECT c FROM Customer c WHERE" +
     "(SELECT SUM(o.totalAmount) FROM Order o WHERE o.customer = c ) > : amount")
-    List<Customer> findTopSpenders(@Param("amount") double amount);
+    List<Customer> findTopSpenders(@Param("amount") BigDecimal amount);
 
     // Native Query = count customers by City
     @Query(value = "SELECT address_city as city, COUNT(*) as count FROM customers GROUP BY address_city",
     nativeQuery = true)
     List<Object[]> countCustomersByCity();
+
 
 }
